src/learn/requirements.txt
redis==4.3.4
kafka-python==2.0.2
torch==2.3.1+cpu
torchvision==0.18.1+cpu